keep nonStandard properties as sub-property
The idea behind the typing was to have a strictly defined object structure with properties according to standard and then to have all non-standard properties accessible on the same level with a default type (a list of `unknown` Property instances).

However, this type is not possible without any drawbacks as of now. It requires bypassing the compiler when constructing calendar structures in the code, resulting in many places where you'll have to use the `as` keyword. This hurts type safety, which is a main goal of this library.

(see TypeScript issue: https://github.com/microsoft/TypeScript/issues/17867)

Additionally, adding new properties later on will not result in a BC-break. Keeping all non-standards on the same level as standards means that accessing `calendar.FOO` is `undefined` or a list of unknown Property instances. The day we add `FOO` as standard component, this type will suddenly change. It might not even be an array anymore.

The other way around, your application will still have to change with new properties being added to the standard, but it won't fail (at least not from a type-safety perspective). Because all non-standard properties can be `undefined` by design, you'll have to check whether the definition exists anyway. A new standard property moves it from the non-standard property, leaving it as `undefined` (which is something your implementation must expected anyway).